Reference

CreditMetrics.CreditMetrics — Module

Implements the CreditMetrics model for portfolio credit risk

CreditMetrics.CheckLevel — Type

Type to indicate the depth of runtime checking

CreditMetrics.Checked — Type

Check for consistency at runtime

CreditMetrics.Unchecked — Type

Do not check for consistency at runtime - I know what I'm doing

CreditMetrics.to_probabilities — Method

to_probabilities(thr, ::Checked)

Execute to_probabilities(thr) but first verify that thr is indeed a proper vector of thresholds.

CreditMetrics.to_probabilities — Method

to_probabilities(thr)

Convert a vector of normal distribution thresholds thr to a vector of probabilities.

CreditMetrics.to_thresholds — Method

to_thresholds(ps)

Convert a vector of probabilities ps to a vector of normal distribution thresholds.

CreditMetrics.to_thresholds — Method

to_thresholds(ps, ::Checked)

Execute to_thresholds(ps) but first verify that ps is a proper vector of probabilities.

CreditMetrics.verify_probabilities — Method

verify_probabilities(ps)

Verify that ps is a vector of probabilities.

CreditMetrics.verify_thresholds — Method

verify_thresholds(thrs)

Verify that thrs is a proper vector of thresholds.