Reference

CreditMetrics.CreditMetrics — Module

Implements the CreditMetrics model for portfolio credit risk

CreditMetrics.AbstractProbabilityVector — Type

Abstract type for probability vectors with entries of type 'T' summing to one

CreditMetrics.AbstractRealLineThresholds — Type

Abstract type for a weakly sorted vector of thresholds of type 'T' where values are unconstrained

CreditMetrics.AbstractThresholdLink — Type

Abstract type for link functions between probabilities and values on the whole real line

CreditMetrics.AbstractUnitIntervalThresholds — Type

Abstract type for a weakly sorted vector of thresholds of type 'T' where values are between zero and one

CreditMetrics.DeformedRealLineThresholds — Type

A weakly sorted vector of values on the real line that has been affinely deformed.

DeformedRealLineThresholds holds the specification of the affine deformation and defers calculation of the deformed thresholds until they are needed.

CreditMetrics.NormalThresholdLink — Type

Normal distribution link specification using StatsFuns implementation of normal cdf and invcdf

CreditMetrics.ProbabilityVector — Type

Vector of probabilities that sum to one

The ProbabilityVector type is used for checking that a vector of probabilities sums to one and using the type system to express that.

CreditMetrics.RealLineThresholds — Type

A weakly sorted vector of values on the real line.

The UnitIntervalThresholds struct is used to check that values are weakly sorted, and using the type system to express that.

CreditMetrics.StudentThresholdLink — Type

Student's distribution link specification using StatsFuns implementation of Student's cdf and invcdf

CreditMetrics.UnitIntervalThresholds — Type

A weakly sorted vector of values on the interval from zero and one.

The UnitIntervalThresholds struct is used to check that values are in interval and weakly sorted, and using the type system to express that.

CreditMetrics.cdf — Method

cdf(atl::AbstractThresholdLink)

Get the link function from real line to probability for the link specification 'atl'

CreditMetrics.cdf — Method

cdf(ntl::NormalThresholdLink)

Get the link function from real line to probability for 'ntl' which is always the normal cdf

CreditMetrics.cdf — Method

cdf(stl::StudentThresholdLink)

Get the link function from real line to probability for 'stl' which is the Student's cdf with degrees of freedom specified by 'stl'

CreditMetrics.invcdf — Method

invcdf(atl::AbstractThresholdLink)

Get the link function from probability to real line for the link specification 'atl'

CreditMetrics.invcdf — Method

invcdf(ntl::NormalThresholdLink)

Get the link function from probability to real line for 'ntl' which is always the normal inverse cdf

CreditMetrics.invcdf — Method

invcdf(stl::StudentThresholdLink)

Get the link function from probability to real line for 'stl' which is the Student's inverse cdf with degrees of freedom specified by 'stl'

CreditMetrics.probabilities — Method

probabilities(apv::AbstractProbabilityVector)

Get the probabilities of 'apv' in a raw format

CreditMetrics.probabilities — Method

probabilities(pv::ProbabilityVector)

Get the probabilities of 'pv' in a raw format

CreditMetrics.thresholds — Method

thresholds(arlt::AbstractRealLineThresholds)

Get the thresholds of 'arlt' in a raw format

CreditMetrics.thresholds — Method

thresholds(auit::AbstractUnitIntervalThresholds)

Get the thresholds of 'auit' in a raw format

CreditMetrics.thresholds — Method

thresholds(rlt::RealLineThresholds)

Extract the raw thresholds from 'rlt'

CreditMetrics.thresholds — Method

thresholds(uit::UnitIntervalThresholds)

Extract raw thresholds from 'uit'

CreditMetrics.thresholds — Method

thresholds(drlt::DeformedRealLineThresholds{T}) where {T}

Calculate the raw thresholds of 'drlt'